Lecture 67 + 68 (read description).
search logic that is aware of relationships; for example when we search React, we want the professors who have the custom field of React to be found. By the end of 67, this was not working. Was not finding professors. End of 68: search query is still not loading Kostas or any professor when React is being searched--